pyproject.toml
[tool.poetry]
name = "todo-fastapi"
version = "0.1.0"
description = ""
authors = ["povarna <povarna@adobe.com>"]
readme = "README.md"
packages = [{include = "todo_fastapi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
tenacity = "^8.4.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"
sqlalchemy = "^2.0.30"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
